# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VaultBench API",
    "version": "0.1.0",
    "description": "Crypto vault and copy-trading performance measured against BTC/ETH/SOL buy-and-hold, on a time-weighted basis with flows removed.\n\nTwo rules govern every response:\n\n1. **Money is a string.** Returns, ratios and money values are exact decimal strings. A JSON number is a float and would reintroduce the error the database's numeric columns exist to avoid.\n2. **Coverage travels with every figure.** Every metrics object carries `coverage` with `daysCovered`, `isFullWindow`, `sampling`, `navQuality` and `headlineEligible`. A 90-day return built from 4 days of downsampled data is reported as such rather than presented as a 90-day return.\n\nClosed and delisted entities are served by default. Excluding them is what makes a leaderboard flattering and wrong."
  },
  "servers": [
    {
      "url": "http://localhost:8787"
    }
  ],
...
  "components": {
    "schemas": {
      "Pagination": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "limit",
          "offset",
          "total"
        ]
      },
      "EntityWithMetrics": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Entity"
          },
          {
            "type": "object",
            "properties": {
              "metrics": {
                "$ref": "#/components/schemas/Metrics"
              }
            },
            "required": [
              "metrics"
            ]
          }
        ]
      },
...
      "Entity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "source": {
            "type": "string",
            "example": "hyperliquid"
          },
          "externalId": {
            "type": "string",
            "example": "0xdfc24b077bc1425ad1dea75bcb6f8158e10df303"
          },
          "kind": {
            "type": "string",
            "example": "vault"
          },
          "name": {
            "type": "string"
          },
          "venue": {
            "type": "string"
          },
          "venueType": {
            "type": "string",
            "example": "dex"
          },
          "marketType": {
            "type": "string",
            "example": "perp"
          },
          "strategyCategory": {
            "type": [
              "string",
              "null"
            ],
            "description": "Hand-assigned. Null means unclassified, not uncategorisable — read it before calling a market-neutral vault an underperformer for trailing BTC.",
            "example": "neutral"
          },
          "baseCurrency": {
            "type": "string"
          },
          "inceptionDate": {
            "type": [
              "string",
              "null"
            ],
            "example": "2026-09-04",
            "description": "UTC date."
          },
          "status": {
            "type": "string",
            "description": "`active`, `closed` or `delisted`. Dead entities are retained deliberately: dropping them is how a leaderboard invents survivorship bias.",
            "example": "active"
          },
          "firstSeenAt": {
            "type": "string"
          },
          "lastSeenAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "externalId",
          "kind",
          "name",
          "venue",
          "venueType",
          "marketType",
          "strategyCategory",
          "baseCurrency",
          "inceptionDate",
          "status",
          "firstSeenAt",
          "lastSeenAt"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "Series": {
        "type": "object",
        "properties": {
          "entityId": {
            "type": "string",
            "format": "uuid"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NavPoint"
            }
          },
          "flows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FlowPoint"
            }
          }
        },
        "required": [
          "entityId",
          "points",
          "flows"
        ]
      },
...
      "Comparison": {
        "type": "object",
        "properties": {
          "entityId": {
            "type": "string",
            "format": "uuid"
          },
          "startAsOf": {
            "type": [
              "string",
              "null"
            ],
            "example": "2026-09-04",
            "description": "UTC date."
          },
          "endAsOf": {
            "type": [
              "string",
              "null"
            ],
            "example": "2026-09-04",
            "description": "UTC date."
          },
          "entity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndexPoint"
            }
          },
          "benchmarks": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/IndexPoint"
              }
            },
            "description": "Keyed by symbol. Each series is aligned to the same start date as the entity."
          },
          "coverage": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Coverage"
              },
              {
                "type": [
                  "object",
                  "null"
                ]
              }
            ]
          },
          "note": {
            "type": "string",
            "description": "Method disclosure that must be shown wherever the chart is shown."
          }
        },
        "required": [
          "entityId",
          "startAsOf",
          "endAsOf",
          "entity",
          "benchmarks",
          "coverage",
          "note"
        ]
      },
...
    },
...
  },
  "paths": {
    "/entities": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "List entities with their metrics for one window",
        "description": "Closed and delisted entities are included by default. Pass `status=active` to exclude them, and understand that doing so reintroduces survivorship bias: the dead entities are the ones a leaderboard never shows you.",
        "parameters": [
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "default": 0,
              "example": 90
            },
            "required": false,
            "name": "window",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "source",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "kind",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "marketType",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "directional",
                "neutral",
                "yield"
              ]
            },
            "required": false,
            "name": "strategyCategory",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "headlineEligible",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "fullWindow",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "twr",
                "alphaBtc",
                "maxDrawdown",
                "volatility",
                "followerGap",
                "name"
              ],
              "default": "twr"
            },
            "required": false,
            "name": "sort",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "required": false,
            "name": "direction",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "default": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Entities and their newest metrics row for the window.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "windowDays": {
                      "type": "integer"
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "entities": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EntityWithMetrics"
                      }
                    }
                  },
                  "required": [
                    "windowDays",
                    "pagination",
                    "entities"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/entities/{id}": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "One entity with every window it has metrics for",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The entity and one metrics row per window.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Entity"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "metrics": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Metrics"
                          }
                        }
                      },
                      "required": [
                        "metrics"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Entity not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/entities/{id}/series": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "The derived per-unit value series, and the flows removed from it",
        "description": "The flows are returned alongside on purpose: they are what was subtracted to stop a deposit reading as a gain, and publishing them is what makes the per-unit series checkable rather than merely asserted.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "to",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Per-unit series and flows.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Series"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/compare": {
      "get": {
        "tags": [
          "compare"
        ],
        "summary": "An entity against buy-and-hold benchmarks, on one index base",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "entity",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "default": "BTC,ETH,SOL",
              "example": "BTC,ETH"
            },
            "required": false,
            "name": "bench",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "default": 0
            },
            "required": false,
            "name": "window",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Aligned, rebased series.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comparison"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/entities/{id}/followers": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "The depositor cross-section and its return distribution",
        "description": "The gap between an entity's return and its median follower's is the number that matters to somebody deciding whether to copy it, and it is the number no venue publishes.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Followers and distribution.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Followers"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}